+++ gems/aws-sdk-cognitoidentity/lib/aws-sdk-cognitoidentity/customizations.rb
@@ -1,3 +1,7 @@
 # frozen_string_literal: true
 
-require 'aws-sdk-cognitoidentity/customizations/cognito_identity_credentials'
+module Aws
+  module CognitoIdentity
+    autoload :CognitoIdentityCredentials, 'aws-sdk-cognitoidentity/customizations/cognito_identity_credentials'
+  end
+end