+++ gems/aws-sdk-dynamodbstreams/lib/aws-sdk-dynamodbstreams/customizations.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
 # utility classes
-require 'aws-sdk-dynamodbstreams/attribute_translator'
-require 'aws-sdk-dynamodbstreams/attribute_value'
+module Aws
+  module DynamoDBStreams
+    autoload :AttributeTranslator, 'aws-sdk-dynamodbstreams/attribute_translator'
+    autoload :AttributeValue, 'aws-sdk-dynamodbstreams/attribute_value'
+  end
+end