+++ gems/aws-sdk-polly/lib/aws-sdk-polly/customizations.rb
@@ -1,4 +1,8 @@
 # frozen_string_literal: true
 
 # utility classes
-require 'aws-sdk-polly/presigner'
+module Aws
+  module Polly
+    autoload :Presigner, 'aws-sdk-polly/presigner'
+  end
+end